Increasing consistency of config dict consuming
