+++ sashimi/hardware/external_trigger/interface.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 
 class AbstractComm:
-    def __init__(self, address):
+    def __init__(self, address=None, **kwargs):
         self.address = address
 
     def trigger_and_receive_duration(self, config) -> Optional[float]:
