+++ sashimi/hardware/external_trigger/mock.py
@@ -2,8 +2,5 @@ from sashimi.hardware.external_trigger.interface import AbstractComm
 
 
 class MockComm(AbstractComm):
-    def __init__(self):
-        super().__init__(address=None)
-
     def trigger_and_receive_duration(self, config):
         return None
